% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{a_to_b}
\alias{a_to_b}
\title{Map between mappings}
\usage{
a_to_b(a, b)
}
\arguments{
\item{a}{Numeric vector with source numbers.}

\item{b}{Data frame with mapping.}
}
\value{
Numeric vector with destination numbers.
}
\description{
Map between mappings
}
