% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{by_df}
\alias{by_df}
\title{Wrapper for \code{by} that returns a data frame}
\usage{
by_df(data, INDICES, FUN, ..., simplify = TRUE, bind_fx = rbind)
}
\arguments{
\item{data}{an \R object, normally a data frame, possibly a matrix.}

\item{INDICES}{a factor or a list of factors, each of length
   \code{nrow(data)}. For the data frame method, \code{INDICES} can also
   be a formula as in the \code{f} argument of the \code{\link[base]{split}}
   method for data frames.}

\item{FUN}{a function to be applied to (usually data-frame) subsets of
   \code{data}.}

\item{...}{
  Arguments passed on to \code{\link[base:by]{base::by}}
  \describe{
    \item{\code{}}{}
  }}

\item{simplify}{logical: see \code{\link[base]{tapply}}.}

\item{bind_fx}{Binding function for the output (e.g., \code{rbind} or \code{cbind}).}
}
\value{
Data frame with the output of \code{by}
}
\description{
Wrapper for \code{by} that returns a data frame
}
