% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day07.R
\name{cards_strength}
\alias{cards_strength}
\title{Calculate the strength of individual cards}
\usage{
cards_strength(x, joker = FALSE)
}
\arguments{
\item{x}{Vector with cards (strings).}

\item{joker}{Flag to indicate if jokers ('J') are scored differently.}
}
\value{
Card scores
}
\description{
Calculate the strength of individual cards
}
