% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day13.R
\name{day13}
\alias{day13}
\alias{f13a}
\alias{f13b}
\alias{example_data_13}
\title{Day 13: Point of Incidence}
\usage{
f13a(x)

f13b(x)

example_data_13(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f13a(x)} returns .... For Part Two,
\code{f13b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2023/day/13}{Point of Incidence}
}
\details{
\strong{Part One}

With your help, the hot springs team locates an appropriate spring which
launches you neatly and precisely up to the edge of \emph{Lava Island}.

There\'s just one problem: you don\'t see any \emph{lava}.

You \emph{do} see a lot of ash and igneous rock; there are even what look
like gray mountains scattered around. After a while, you make your way
to a nearby cluster of mountains only to discover that the valley
between them is completely full of large \emph{mirrors}. Most of the mirrors
seem to be aligned in a consistent way; perhaps you should head in that
direction?

As you move through the valley of mirrors, you find that several of them
have fallen from the large metal frames keeping them in place. The
mirrors are extremely flat and shiny, and many of the fallen mirrors
have lodged into the ash at strange angles. Because the terrain is all
one color, it\'s hard to tell where it\'s safe to walk or where you\'re
about to run into a mirror.

You note down the patterns of ash (\code{.}) and rocks (\verb{#}) that you see as
you walk (your puzzle input); perhaps by carefully analyzing these
patterns, you can figure out where the mirrors are!

For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#.##..##.
..#.##.#.
##......#
##......#
..#.##.#.
..##..##.
#.#.##.#.

#...##..#
#....#..#
..##..###
#####.##.
#####.##.
..##..###
#....#..#
}\if{html}{\out{</div>}}

To find the reflection in each pattern, you need to find a perfect
reflection across either a horizontal line between two rows or across a
vertical line between two columns.

In the first pattern, the reflection is across a vertical line between
two columns; arrows on each of the two columns point at the line between
the columns:

\if{html}{\out{<div class="sourceCode">}}\preformatted{123456789
    ><
#.##..##.
..#.##.#.
##......#
##......#
..#.##.#.
..##..##.
#.#.##.#.
    ><
123456789
}\if{html}{\out{</div>}}

In this pattern, the line of reflection is the vertical line between
columns 5 and 6. Because the vertical line is not perfectly in the
middle of the pattern, part of the pattern (column 1) has nowhere to
reflect onto and can be ignored; every other column has a reflected
column within the pattern and must match exactly: column 2 matches
column 9, column 3 matches 8, 4 matches 7, and 5 matches 6.

The second pattern reflects across a horizontal line instead:

\if{html}{\out{<div class="sourceCode">}}\preformatted{1 #...##..# 1
2 #....#..# 2
3 ..##..### 3
4v#####.##.v4
5^#####.##.^5
6 ..##..### 6
7 #....#..# 7
}\if{html}{\out{</div>}}

This pattern reflects across the horizontal line between rows 4 and 5.
Row 1 would reflect with a hypothetical row 8, but since that\'s not in
the pattern, row 1 doesn\'t need to match anything. The remaining rows
match: row 2 matches row 7, row 3 matches row 6, and row 4 matches row
5.

To \emph{summarize} your pattern notes, add up \emph{the number of columns} to the
left of each vertical line of reflection; to that, also add \emph{100
multiplied by the number of rows} above each horizontal line of
reflection. In the above example, the first pattern\'s vertical line has
\code{5} columns to its left and the second pattern\'s horizontal line has
\code{4} rows above it, a total of \emph{\code{405}}.

Find the line of reflection in each of the patterns in your notes. \emph{What
number do you get after summarizing all of your notes?}

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
f13a(example_data_13())
f13b(example_data_13())
}
