% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day15.R
\name{day15}
\alias{day15}
\alias{f15a}
\alias{f15b}
\alias{example_data_15}
\title{Day 15: Lens Library}
\usage{
f15a(x)

f15b(x)

example_data_15(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f15a(x)} returns .... For Part Two,
\code{f15b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2023/day/15}{Lens Library}
}
\details{
\strong{Part One}

The newly-focused parabolic reflector dish is sending all of the
collected light to a point on the side of yet another mountain - the
largest mountain on Lava Island. As you approach the mountain, you find
that the light is being collected by the wall of a large facility
embedded in the mountainside.

You find a door under a large sign that says \"Lava Production
Facility\" and next to a smaller sign that says \"Danger - Personal
Protective Equipment required beyond this point\".

As you step inside, you are immediately greeted by a somewhat panicked
[reindeer]{title="do you like my hard hat"} wearing goggles and a
loose-fitting \href{https://en.wikipedia.org/wiki/Hard_hat}{hard hat}{target="_blank"}. The
reindeer leads you to a shelf of goggles and hard hats (you quickly find
some that fit) and then further into the facility. At one point, you
pass a button with a faint snout mark and the label \"PUSH FOR HELP\".
No wonder you were loaded into that \href{1}{trebuchet} so quickly!

You pass through a final set of doors surrounded with even more warning
signs and into what must be the room that collects all of the light from
outside. As you admire the large assortment of lenses available to
further focus the light, the reindeer brings you a book titled
\"Initialization Manual\".

\"Hello!\", the book cheerfully begins, apparently unaware of the
concerned reindeer reading over your shoulder. \"This procedure will let
you bring the Lava Production Facility online - all without burning or
melting anything unintended!\"

\"Before you begin, please be prepared to use the Holiday ASCII String
Helper algorithm (appendix 1A).\" You turn to appendix 1A. The reindeer
leans closer with interest.

The HASH algorithm is a way to turn any
\href{https://en.wikipedia.org/wiki/String_(computer_science)}{string}{target="_blank"}
of characters into a single \emph{number} in the range 0 to 255. To run the
HASH algorithm on a string, start with a \emph{current value} of \code{0}. Then,
for each character in the string starting from the beginning:
\itemize{
\item Determine the \href{https://en.wikipedia.org/wiki/ASCII#Printable_characters}{ASCII code}{target="_blank"}
for the current character of the string.
\item Increase the \emph{current value} by the ASCII code you just determined.
\item Set the \emph{current value} to itself multiplied by \code{17}.
\item Set the \emph{current value} to the
\href{https://en.wikipedia.org/wiki/Modulo}{remainder}{target="_blank"}
of dividing itself by \code{256}.
}

After following these steps for each character in the string in order,
the \emph{current value} is the output of the HASH algorithm.

So, to find the result of running the HASH algorithm on the string
\code{HASH}:
\itemize{
\item The \emph{current value} starts at \code{0}.
\item The first character is \code{H}; its ASCII code is \code{72}.
\item The \emph{current value} increases to \code{72}.
\item The \emph{current value} is multiplied by \code{17} to become \code{1224}.
\item The \emph{current value} becomes \emph{\code{200}} (the remainder of \code{1224} divided
by \code{256}).
\item The next character is \code{A}; its ASCII code is \code{65}.
\item The \emph{current value} increases to \code{265}.
\item The \emph{current value} is multiplied by \code{17} to become \code{4505}.
\item The \emph{current value} becomes \emph{\code{153}} (the remainder of \code{4505} divided
by \code{256}).
\item The next character is \code{S}; its ASCII code is \code{83}.
\item The \emph{current value} increases to \code{236}.
\item The \emph{current value} is multiplied by \code{17} to become \code{4012}.
\item The \emph{current value} becomes \emph{\code{172}} (the remainder of \code{4012} divided
by \code{256}).
\item The next character is \code{H}; its ASCII code is \code{72}.
\item The \emph{current value} increases to \code{244}.
\item The \emph{current value} is multiplied by \code{17} to become \code{4148}.
\item The \emph{current value} becomes \emph{\code{52}} (the remainder of \code{4148} divided
by \code{256}).
}

So, the result of running the HASH algorithm on the string \code{HASH} is
\emph{\code{52}}.

The \emph{initialization sequence} (your puzzle input) is a comma-separated
list of steps to start the Lava Production Facility. \emph{Ignore newline
characters} when parsing the initialization sequence. To verify that
your HASH algorithm is working, the book offers the sum of the result of
running the HASH algorithm on each step in the initialization sequence.

For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{rn=1,cm-,qp=3,cm=2,qp-,pc=4,ot=9,ab=5,pc-,pc=6,ot=7
}\if{html}{\out{</div>}}

This initialization sequence specifies 11 individual steps; the result
of running the HASH algorithm on each of the steps is as follows:
\itemize{
\item \code{rn=1} becomes \emph{\code{30}}.
\item \verb{cm-} becomes \emph{\code{253}}.
\item \code{qp=3} becomes \emph{\code{97}}.
\item \code{cm=2} becomes \emph{\code{47}}.
\item \verb{qp-} becomes \emph{\code{14}}.
\item \code{pc=4} becomes \emph{\code{180}}.
\item \code{ot=9} becomes \emph{\code{9}}.
\item \code{ab=5} becomes \emph{\code{197}}.
\item \verb{pc-} becomes \emph{\code{48}}.
\item \code{pc=6} becomes \emph{\code{214}}.
\item \code{ot=7} becomes \emph{\code{231}}.
}

In this example, the sum of these results is \emph{\code{1320}}. Unfortunately,
the reindeer has stolen the page containing the expected verification
number and is currently running around the facility with it excitedly.

Run the HASH algorithm on each step in the initialization sequence.
\emph{What is the sum of the results?} (The initialization sequence is one
long line; be careful when copy-pasting it.)

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
f15a(example_data_15())
f15b(example_data_15())
}
