% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day09.R
\name{extrapolate_values}
\alias{extrapolate_values}
\title{Extrapolate values from sequence, see
\href{https://adventofcode.com/2023/day/9}{day 9 - 2023}}
\usage{
extrapolate_values(x, reverse = FALSE)
}
\arguments{
\item{x}{List with parsed histories (rows of numbers).}

\item{reverse}{Boolean flag to indicate if the history should be flipped.}
}
\value{
Numeric value with the predicted value.
}
\description{
Extrapolate values from sequence, see
\href{https://adventofcode.com/2023/day/9}{day 9 - 2023}
}
