% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day01.R
\name{f01_helper}
\alias{f01_helper}
\title{Combines two vectors with digits and returns the sum}
\usage{
f01_helper(first_digit, last_digit)
}
\arguments{
\item{first_digit}{Vector with first digits.}

\item{last_digit}{Vector with last digits.}
}
\value{
Sum of the combination of both vectors
}
\description{
Combines two vectors with digits and returns the sum
}
\examples{
f01_helper("1", "2") # should return 12
}
