% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day15.R
\name{find_label_lens}
\alias{find_label_lens}
\title{Find label-lens in box, see
\href{https://adventofcode.com/2023/day/15}{day 15 - 2023}}
\usage{
find_label_lens(contents, label)
}
\arguments{
\item{contents}{String with contents of the box.}

\item{label}{String with box label.}
}
\value{
Boolean value to indicate if the label-lens were found.
}
\description{
Find label-lens in box, see
\href{https://adventofcode.com/2023/day/15}{day 15 - 2023}
}
