% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day13.R
\name{find_mid_point_reflection}
\alias{find_mid_point_reflection}
\title{Find mid point for the reflection, see
\href{https://adventofcode.com/2023/day/13}{day 13 - 2023}}
\usage{
find_mid_point_reflection(x, diff_count = 0)
}
\arguments{
\item{x}{Matrix with pattern.}

\item{diff_count}{Number of differences allowed.}
}
\value{
Data frame with score for each pattern.
}
\description{
Find mid point for the reflection, see
\href{https://adventofcode.com/2023/day/13}{day 13 - 2023}
}
