% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{find_min_location}
\alias{find_min_location}
\title{Find minimum locations for seed ranges}
\usage{
find_min_location(x, start, length)
}
\arguments{
\item{x}{String with Almanac.}

\item{start}{Starting point of the seed range.}

\item{length}{Length of the seed range.}
}
\value{
Minimum location for the given range.
}
\description{
Find minimum locations for seed ranges. Adapted from
\url{https://www.reddit.com/r/adventofcode/comments/18buwiz/comment/kc78ou6/}
}
