% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{find_minimum_set}
\alias{find_minimum_set}
\title{Find the minimum set of cubes for a game, see
\href{https://adventofcode.com/2023/day/2/#part2}{day 2 - 2023}}
\usage{
find_minimum_set(x)
}
\arguments{
\item{x}{Data frame with game details.}
}
\value{
Data frame with minim set of cubes required.
}
\description{
Find the minimum set of cubes for a game, see
\href{https://adventofcode.com/2023/day/2/#part2}{day 2 - 2023}
}
\examples{
demo <- parse_game("Game 1: 3 blue, 4 red; 1 red, 2 green, 6 blue; 2 green")
find_minimum_set(demo)
}
