% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{first_true_element}
\alias{first_true_element}
\title{Get the first \code{TRUE} consecutive element from reference point}
\usage{
first_true_element(x, i)
}
\arguments{
\item{x}{Vector with logical values.}

\item{i}{Reference point}
}
\value{
Numeric value with the index for the first \code{TRUE} consecutive element
}
\description{
Get the first \code{TRUE} consecutive element from reference point
}
