% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day15.R
\name{focus_power}
\alias{focus_power}
\title{Focus lens power, see
\href{https://adventofcode.com/2023/day/15}{day 15 - 2023}}
\usage{
focus_power(x)
}
\arguments{
\item{x}{Hash table with mapped boxes.}
}
\value{
Data frame with labels for boxes and power value.
}
\description{
Focus lens power, see
\href{https://adventofcode.com/2023/day/15}{day 15 - 2023}
}
