% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_adjacent_elements}
\alias{get_adjacent_elements}
\title{Get adjacent elements from a 2D object}
\usage{
get_adjacent_elements(x, i, j)
}
\arguments{
\item{x}{2D object.}

\item{i}{Reference row.}

\item{j}{Reference column.}
}
\value{
Matrix with adjacent elements
}
\description{
Get adjacent elements from a 2D object
}
