% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_key}
\alias{get_key}
\title{Get hashing key from matrix elements}
\usage{
get_key(x)
}
\arguments{
\item{x}{Matrix.}
}
\value{
String with hashing key for matrix.
}
\description{
Get hashing key from matrix elements
}
