% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_linear_subscripts}
\alias{get_linear_subscripts}
\title{Convert 2D subscript to linear}
\usage{
get_linear_subscripts(x, i, j)
}
\arguments{
\item{x}{2D object.}

\item{i}{Reference row.}

\item{j}{Reference column.}
}
\value{
Numeric value with linear index.
}
\description{
Convert 2D subscript to linear
}
