% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day15.R
\name{get_mapping}
\alias{get_mapping}
\title{Get mapping for string, see
\href{https://adventofcode.com/2023/day/15}{day 15 - 2023}}
\usage{
get_mapping(x, offset = 0, max = 256, mult = 17)
}
\arguments{
\item{x}{String to be hashed.}

\item{offset}{Integer with hashing offset.}

\item{max}{Maximum number of elements in hashing table.}

\item{mult}{Integer with scaling factor.}
}
\value{
Integer with mapped value.
}
\description{
Get mapping for string, see
\href{https://adventofcode.com/2023/day/15}{day 15 - 2023}
}
