% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_matrix_subscripts}
\alias{get_matrix_subscripts}
\title{Convert linear subscript to 2D}
\usage{
get_matrix_subscripts(x, ind)
}
\arguments{
\item{x}{2D object.}

\item{ind}{Numeric value with linear index}
}
\value{
2D array with row and column indices.
}
\description{
Convert linear subscript to 2D
}
