% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{get_next_tiles}
\alias{get_next_tiles}
\title{Get next field tiles, see
\href{https://adventofcode.com/2023/day/10}{day 10 - 2023}}
\usage{
get_next_tiles(env, chr, pos, all = FALSE)
}
\arguments{
\item{env}{Environment with the current field, \code{x}.}

\item{chr}{Character with current position's element.}

\item{pos}{Data frame with current position (\code{i}, \code{j}).}

\item{all}{Boolean flag to indicate if all tiles (including visited)
should be returned.}
}
\value{
Data frame with next tile positions.
}
\description{
Get next field tiles, see
\href{https://adventofcode.com/2023/day/10}{day 10 - 2023}
}
