% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%gcd\%}
\alias{\%gcd\%}
\title{Greatest Common Divisor (GCD)}
\source{
https://rosettacode.org/wiki/Greatest_common_divisor#R
}
\usage{
u \%gcd\% v
}
\arguments{
\item{u}{Numeric value.}

\item{v}{Numeric value.}
}
\value{
Numeric value with GCD for u and v.
}
\description{
Greatest Common Divisor (GCD)
}
