% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day07.R
\name{hand_strength}
\alias{hand_strength}
\title{Calculate the strength of a group of cards / a hand, see
\href{https://adventofcode.com/2023/day/7}{day 7 - 2023}}
\usage{
hand_strength(x, joker = FALSE)
}
\arguments{
\item{x}{Vector with cards (strings).}

\item{joker}{Flag to indicate if jokers ('J') are scored differently.}
}
\value{
Hand score
}
\description{
Calculate the strength of a group of cards / a hand, see
\href{https://adventofcode.com/2023/day/7}{day 7 - 2023}
}
