% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hashkeys}
\alias{hashkeys}
\title{Get all hashkeys of a hash table}
\usage{
hashkeys(h)
}
\arguments{
\item{h}{Hashtable, created with \code{utils::hashtab}.}
}
\value{
List of keys.
}
\description{
Get all hashkeys of a hash table
}
