% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ind2sub}
\alias{ind2sub}
\title{Convert a linear subscript to matrix subscripts}
\usage{
ind2sub(x, ind)
}
\arguments{
\item{x}{2D object.}

\item{ind}{Numeric value with linear subscript.}
}
\value{
2D array with row and column subscripts.
}
\description{
Convert a linear subscript to matrix subscripts
}
