% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lapply_df}
\alias{lapply_df}
\title{Wrapper for \code{lapply} that returns a data frame}
\usage{
lapply_df(X, FUN, ..., bind_fx = rbind)
}
\arguments{
\item{X}{a vector (atomic or list) or an \code{\link[base]{expression}}
    object.  Other objects (including classed objects) will be coerced
    by \code{base::\link[base]{as.list}}.}

\item{FUN}{the function to be applied to each element of \code{X}:
    see \sQuote{Details}.  In the case of functions like
    \code{+}, \code{\%*\%}, the function name must be backquoted or quoted.}

\item{...}{
  Arguments passed on to \code{\link[base:lapply]{base::lapply}}
  \describe{
    \item{\code{}}{}
  }}

\item{bind_fx}{Binding function for the output (e.g., \code{rbind} or \code{cbind}).}
}
\value{
Data frame with the output of \code{lapply}
}
\description{
Wrapper for \code{lapply} that returns a data frame
}
