% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last_true_element}
\alias{last_true_element}
\title{Get the last \code{TRUE} consecutive element from reference point}
\usage{
last_true_element(x, i)
}
\arguments{
\item{x}{Vector with logical values.}

\item{i}{Reference point}
}
\value{
Numeric value with the index for the last \code{TRUE} consecutive element
}
\description{
Get the last \code{TRUE} consecutive element from reference point
}
