% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lcm}
\alias{lcm}
\title{Least Common Multiple (LCM)}
\usage{
lcm(u, v)
}
\arguments{
\item{u}{Numeric value.}

\item{v}{Numeric value.}
}
\value{
Numeric value with LCM for u and v.
}
\description{
Least Common Multiple (LCM)
}
