% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day15.R
\name{map_boxes}
\alias{map_boxes}
\title{Map sequence of boxes , see
\href{https://adventofcode.com/2023/day/15}{day 15 - 2023}}
\usage{
map_boxes(x)
}
\arguments{
\item{x}{List with parsed sequence.}
}
\value{
Hash table with mapped boxes.
}
\description{
Map sequence of boxes , see
\href{https://adventofcode.com/2023/day/15}{day 15 - 2023}
}
