% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{map_schematic}
\alias{map_schematic}
\title{Map schematic to extract numeric values adjacent to symbol, see
\href{https://adventofcode.com/2023/day/3}{day 3 - 2023}}
\usage{
map_schematic(x)
}
\arguments{
\item{x}{Data frame with parsed schematic.}
}
\value{
Data frame with part numbers.
}
\description{
Map schematic to extract numeric values adjacent to symbol, see
\href{https://adventofcode.com/2023/day/3}{day 3 - 2023}
}
