% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{map_schematic_gears}
\alias{map_schematic_gears}
\title{Map schematic to extract numeric values adjacent to the gear symbol, \code{*},
see \href{https://adventofcode.com/2023/day/3#part2}{day 3 - 2023}}
\usage{
map_schematic_gears(x)
}
\arguments{
\item{x}{Data frame with parsed schematic.}
}
\value{
Data frame with gear ratios.
}
\description{
Map schematic to extract numeric values adjacent to the gear symbol, \code{*},
see \href{https://adventofcode.com/2023/day/3#part2}{day 3 - 2023}
}
