% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{navigate_pipe}
\alias{navigate_pipe}
\title{Navigate field of pipes, see
\href{https://adventofcode.com/2023/day/10}{day 10 - 2023}}
\usage{
navigate_pipe(x)
}
\arguments{
\item{x}{Matrix with field of pipes.}
}
\value{
List with visited pipes to form a loop from the starting point.
}
\description{
Navigate field of pipes, see
\href{https://adventofcode.com/2023/day/10}{day 10 - 2023}
}
