% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day15.R
\name{new_box}
\alias{new_box}
\title{Create new box, see
\href{https://adventofcode.com/2023/day/15}{day 15 - 2023}}
\usage{
new_box(label, lens)
}
\arguments{
\item{label}{String with box label.}

\item{lens}{String with box lens.}
}
\value{
String with new box.
}
\description{
Create new box, see
\href{https://adventofcode.com/2023/day/15}{day 15 - 2023}
}
