% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{parse_cubes}
\alias{parse_cubes}
\title{Parse string with game details, see
\href{https://adventofcode.com/2023/day/2}{day 2 - 2023}}
\usage{
parse_cubes(x)
}
\arguments{
\item{x}{String with game description.}
}
\value{
Data frame with game details (cubes)
}
\description{
Parse string with game details, see
\href{https://adventofcode.com/2023/day/2}{day 2 - 2023}
}
\examples{
parse_cubes("3 blue, 4 red")
}
