% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{parse_game}
\alias{parse_game}
\title{Parse string with game details, see
\href{https://adventofcode.com/2023/day/2}{day 2 - 2023}}
\usage{
parse_game(x)
}
\arguments{
\item{x}{String with game description.}
}
\value{
Data frame with game details.
}
\description{
Parse string with game details, see
\href{https://adventofcode.com/2023/day/2}{day 2 - 2023}
}
\examples{
parse_game("Game 1: 3 blue, 4 red; 1 red, 2 green, 6 blue; 2 green")
}
