% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{parse_map}
\alias{parse_map}
\title{Parse mapping from Almanac}
\usage{
parse_map(x, header = NULL)
}
\arguments{
\item{x}{String with Almanac.}

\item{header}{String with the header of the mapping in the Almanac.}
}
\value{
Data frame with mapping.
}
\description{
Parse mapping from Almanac
}
