% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day08.R
\name{parse_network}
\alias{parse_network}
\title{Parse network map, see
\href{https://adventofcode.com/2023/day/8}{day 8 - 2023}}
\usage{
parse_network(x)
}
\arguments{
\item{x}{String with network map.}
}
\value{
Data frame with parsed network map.
}
\description{
Parse network map, see
\href{https://adventofcode.com/2023/day/8}{day 8 - 2023}
}
