% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_numbers}
\alias{parse_numbers}
\title{Parse numbers from text section}
\usage{
parse_numbers(x, header = "")
}
\arguments{
\item{x}{String of characters with target text.}

\item{header}{String with the header of the target section.}
}
\value{
Numeric vector.
}
\description{
Parse numbers from text section
}
\examples{
parse_numbers("Distance:  9  40  200", "Distance:")
}
