% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day13.R
\name{parse_patterns}
\alias{parse_patterns}
\title{Parse patterns, see
\href{https://adventofcode.com/2023/day/13}{day 13 - 2023}}
\usage{
parse_patterns(x)
}
\arguments{
\item{x}{String with patterns.}
}
\value{
List with 2D matrices for each pattern.
}
\description{
Parse patterns, see
\href{https://adventofcode.com/2023/day/13}{day 13 - 2023}
}
