% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{parse_schematic}
\alias{parse_schematic}
\title{Parse schematic string, see
\href{https://adventofcode.com/2023/day/3}{day 3 - 2023}}
\usage{
parse_schematic(x)
}
\arguments{
\item{x}{String with schematic.}
}
\value{
Data frame with schematic
}
\description{
Parse schematic string, see
\href{https://adventofcode.com/2023/day/3}{day 3 - 2023}
}
\examples{
parse_schematic("467..114..")
}
