% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{parse_seed_ranges}
\alias{parse_seed_ranges}
\title{Parse seed ranges from Almanac}
\usage{
parse_seed_ranges(x)
}
\arguments{
\item{x}{String with Almanac.}
}
\value{
Data frame with seed ranges.
}
\description{
Parse seed ranges from Almanac
}
