% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{parse_seeds}
\alias{parse_seeds}
\title{Parse seeds from Almanac}
\usage{
parse_seeds(x)
}
\arguments{
\item{x}{String with Almanac.}
}
\value{
Numeric vector with the seeds.
}
\description{
Parse seeds from Almanac
}
