% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day15.R
\name{remove_label_lens}
\alias{remove_label_lens}
\title{Remove label-lens from box, see
\href{https://adventofcode.com/2023/day/15}{day 15 - 2023}}
\usage{
remove_label_lens(contents, label)
}
\arguments{
\item{contents}{String with contents of the box.}

\item{label}{String with box label.}
}
\value{
String with new contents of the box.
}
\description{
Remove label-lens from box, see
\href{https://adventofcode.com/2023/day/15}{day 15 - 2023}
}
