% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{replace_chrs}
\alias{replace_chrs}
\title{Replace characters}
\usage{
replace_chrs(x, chrs = "\\\\.", replacement = " ")
}
\arguments{
\item{x}{String.}

\item{chrs}{Vector with characters to be replaced.}

\item{replacement}{String with the replacement string.}
}
\value{
New string with characters replaced.
}
\description{
Replace characters
}
\examples{
replace_chrs("467..114..")
}
