% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day15.R
\name{replace_label_lens}
\alias{replace_label_lens}
\title{Replace label-lens from box, see
\href{https://adventofcode.com/2023/day/15}{day 15 - 2023}}
\usage{
replace_label_lens(contents, label, lens)
}
\arguments{
\item{contents}{String with contents of the box.}

\item{label}{String with box label.}

\item{lens}{String with box lens.}
}
\value{
String with new contents of the box.
}
\description{
Replace label-lens from box, see
\href{https://adventofcode.com/2023/day/15}{day 15 - 2023}
}
