% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day01.R
\name{reverse_string}
\alias{reverse_string}
\title{Reverse a string}
\usage{
reverse_string(x)
}
\arguments{
\item{x}{Original string.}
}
\value{
Reversed string
}
\description{
Reverse a string
}
\examples{
reverse_string("ouch")
}
