% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rotate_rev}
\alias{rotate_rev}
\title{Rotate 2D array 90 deg anti-clockwise}
\usage{
rotate_rev(x)
}
\arguments{
\item{x}{Input array.}
}
\value{
Rotated array
}
\description{
Rotate 2D array 90 deg anti-clockwise
}
