% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shoelace}
\alias{shoelace}
\title{Shoelace formula}
\usage{
shoelace(x, y)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}
}
\value{
Area of polygon enclosed by the given points (\code{x}, \code{y}).
}
\description{
Shoelace formula to find the area of the polygon given by the
vertices found: https://en.wikipedia.org/wiki/Shoelace_formula
}
\examples{
shoelace(x = c(1, 1, 4, 4, 8), y = c(1, 4, 1, 4, 2))
}
