% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day07.R
\name{split_deck}
\alias{split_deck}
\title{Split string with hands of cards, see
\href{https://adventofcode.com/2023/day/7}{day 7 - 2023}}
\usage{
split_deck(x)
}
\arguments{
\item{x}{String with hands of cards.}
}
\value{
Data frame with hands
}
\description{
Split string with hands of cards, see
\href{https://adventofcode.com/2023/day/7}{day 7 - 2023}
}
