% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day07.R
\name{split_hand}
\alias{split_hand}
\title{Split string with cards, see
\href{https://adventofcode.com/2023/day/7}{day 7 - 2023}}
\usage{
split_hand(x)
}
\arguments{
\item{x}{String with cards.}
}
\value{
Vector of strings for each card
}
\description{
Split string with cards, see
\href{https://adventofcode.com/2023/day/7}{day 7 - 2023}
}
