% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sub2ind}
\alias{sub2ind}
\title{Convert matrix subscripts to a linear subscript}
\usage{
sub2ind(x, i, j)
}
\arguments{
\item{x}{2D object.}

\item{i}{Reference row.}

\item{j}{Reference column.}
}
\value{
Numeric value with linear subscript.
}
\description{
Convert matrix subscripts to a linear subscript
}
