% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day14.R
\name{tilt_cycle}
\alias{tilt_cycle}
\title{Tilting cycle, see
\href{https://adventofcode.com/2023/day/14}{day 14 - 2023}}
\usage{
tilt_cycle(x)
}
\arguments{
\item{x}{Parsed platform, matrix.}
}
\value{
Tilted platform, matrix.
}
\description{
Tilting cycle, see
\href{https://adventofcode.com/2023/day/14}{day 14 - 2023}
}
