% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day14.R
\name{tilt_platform}
\alias{tilt_platform}
\title{Tilt platform, see
\href{https://adventofcode.com/2023/day/14}{day 14 - 2023}}
\usage{
tilt_platform(x, direction = 1)
}
\arguments{
\item{x}{Parsed platform, matrix.}

\item{direction}{Integer with tilting direction: 1 (North), 2 (West),
3 (South) and 4 (East).}
}
\value{
Tilted platform, matrix.
}
\description{
Tilt platform, see
\href{https://adventofcode.com/2023/day/14}{day 14 - 2023}
}
