% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day14.R
\name{total_load}
\alias{total_load}
\title{Calculate total load, see
\href{https://adventofcode.com/2023/day/14}{day 14 - 2023}}
\usage{
total_load(x)
}
\arguments{
\item{x}{Tilted platform, matrix.}
}
\value{
Total load.
}
\description{
Calculate total load, see
\href{https://adventofcode.com/2023/day/14}{day 14 - 2023}
}
