% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day08.R
\name{traverse_map}
\alias{traverse_map}
\title{Traverse network map, see
\href{https://adventofcode.com/2023/day/8}{day 8 - 2023}}
\usage{
traverse_map(route, parsed_network, start, target)
}
\arguments{
\item{route}{String with route (e.g., 'LR').}

\item{parsed_network}{Data frame with parsed network map.}

\item{start}{String with starting point.}

\item{target}{Vector of strings with targets/end points.}
}
\value{
List with various elements resulting from traversing the network.
}
\description{
Traverse network map, see
\href{https://adventofcode.com/2023/day/8}{day 8 - 2023}
}
