% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{validate_hands}
\alias{validate_hands}
\title{Validate hand based on game configuration, see
\href{https://adventofcode.com/2023/day/2}{day 2 - 2023}}
\usage{
validate_hands(x, limits_df)
}
\arguments{
\item{x}{Data frame with game details.}

\item{limits_df}{Data frame with game configuration.}
}
\value{
Vector of \code{TRUE} and \code{FALSE} values linked to each hand
}
\description{
Validate hand based on game configuration, see
\href{https://adventofcode.com/2023/day/2}{day 2 - 2023}
}
\examples{
demo <- parse_game("Game 1: 3 blue, 4 red; 1 red, 2 green, 6 blue; 2 green")
limits_df <- data.frame(
  colour = c("red", "green", "blue"),
  max_count = c(12, 13, 14)
)
validate_hands(demo, limits_df)
}
