% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\name{winning_options}
\alias{winning_options}
\title{Compute winning options}
\usage{
winning_options(time, distance)
}
\arguments{
\item{time}{Numeric value with time.}

\item{distance}{Numeric value with distance.}
}
\value{
Sum of the winning options.
}
\description{
Compute winning options
}
